update to nav bar
